Initial Commit
This was as far as Greg and I got in class. There was a slight issue
with our logic in checking for the length of the scores array. We
created an array of length 22 that could hold up to 21 scores. In
bowling there could be either 20 or 21 scores depending on whether the
player earned an extra roll. We checked the predetermined length of an
array in order to determine how many scores there are. This is a
mistake. To fix this mistake, we will initialize an array of all -1
values and check if the 21st value has been replaced with a real value.
This is a slightly better approach.